Made runnable not functions
when we run the files, and we run sleepindex.py, we will run the script and not certain functions. So I am just running full-csv-second. off-wrist is not a value we find ourselves. If u need it then look for it in the file urself and assume that it is ON wrist.